+++ packages/react-components/src/checkbox/index.js
@@ -4,32 +4,67 @@ import styled from 'styled-components'
 import { P1 } from '../text/paragraph'
 import { colorGrayscale } from '@twreporter/core/lib/constants/color'
 
+const disableColor = colorGrayscale.gray400
+const activeColor = colorGrayscale.gray800
+
 const Container = styled.div`
   display: flex;
   flex-direction: 'row';
 `
 
 const Input = styled.input`
+  width: 0;
+  height: 0;
+  opacity: 0;
+  z-index: -1;
+`
+
+const Label = styled.label`
+  display: block;
+  position: relative;
+  padding-left: 25px;
+`
+
+const Indicator = styled.div`
   width: 16px;
   height: 16px;
-  border: solid 1px;
+  background: ${props =>
+    props.value
+      ? props.disabled
+        ? disableColor
+        : activeColor
+      : 'rgba(0, 0, 0, 0)'};
+  position: absolute;
+  top: 0px;
+  left: 0px;
+  border-color: ${props => (props.disabled ? disableColor : activeColor)};
   border-radius: 2px;
-  accent-color: ${colorGrayscale.gray800};
-  margin-left: 0px;
-  margin-right: 8px;
+  box-shadow: 0 0 0 1px
+    ${props => (props.disabled ? disableColor : activeColor)};
+  margin: 3px 8px 0px 0px;
 
-  &:enabled:not(:checked) {
-    border-color: ${colorGrayscale.gray800};
-    background: rgba(0, 0, 0, 0);
-    -webkit-appearance: none;
-    -moz-appearance: none;
-    appearance: none;
+  &::after {
+    content: '';
+    position: absolute;
+    display: none;
+  }
+
+  ${Input}:checked + &::after {
+    display: block;
+    left: 5px;
+    top: 0px;
+    width: 4px;
+    height: 10px;
+    border: solid white;
+    border-width: 0 2px 2px 0;
+    -webkit-transform: rotate(45deg);
+    -ms-transform: rotate(45deg);
+    transform: rotate(45deg);
   }
 `
 
 const ColorP1 = styled(P1)`
-  color: ${props =>
-    props.disabled ? colorGrayscale.gray400 : colorGrayscale.gray800};
+  color: ${props => (props.disabled ? disableColor : activeColor)};
 `
 
 export const Checkbox = ({
@@ -44,13 +79,16 @@ export const Checkbox = ({
   }
   return (
     <Container {...props}>
-      <Input
-        type="checkbox"
-        checked={value}
-        disabled={disabled}
-        onChange={handleChange}
-      />
-      {label && <ColorP1 text={label} disabled={disabled} />}
+      <Label>
+        {label && <ColorP1 text={label} disabled={disabled} />}
+        <Input
+          type="checkbox"
+          checked={value}
+          disabled={disabled}
+          onChange={handleChange}
+        />
+        <Indicator value={value} disabled={disabled} />
+      </Label>
     </Container>
   )
 }