+++ packages/react-components/src/listing-page/components/card-list.js
@@ -6,6 +6,7 @@ import { ArticleCard } from '../../card'
 import FetchingWrapper from '../../is-fetching-wrapper'
 import Divider from '../../divider'
 import Link from '../../customized-link'
+import { DesktopOnly, MobileOnly } from '../../rwd'
 // constants
 import mockup from '../constants/mockup-spec'
 // @twreporter
@@ -27,17 +28,8 @@ const _ = {
   map,
 }
 
-const DesktopOnly = styled.div`
-  display: none;
-  ${mq.desktopAndAbove`
-    display: flex;
-  `}
-`
-const MobileOnly = styled.div`
-  display: none;
-  ${mq.tabletAndBelow`
-    display: flex;
-  `}
+const Card = styled(ArticleCard)`
+  width: 100%;
 `
 const Container = styled.div`
   display: flex;
@@ -137,10 +129,10 @@ const CardList = ({
       <Item key={id}>
         <Link {...link}>
           <DesktopOnly>
-            <ArticleCard {...articleCardProps} size={ArticleCard.Size.L} />
+            <Card {...articleCardProps} size={ArticleCard.Size.L} />
           </DesktopOnly>
           <MobileOnly>
-            <ArticleCard {...articleCardProps} size={ArticleCard.Size.S} />
+            <Card {...articleCardProps} size={ArticleCard.Size.S} />
           </MobileOnly>
           <StyledDivider />
         </Link>